Adding alerts system and starting the save system

+++ Assets/Editor/BuildScripts/BuildUtils.cs
@@ -6,7 +6,6 @@ public static class BuildUtils
 {
     private const string projectName = "TaleTaleEngine.exe";
 
-    // Common method to handle the build process
     public static void BuildProject(string[] scenes, string buildPath, BuildTarget buildTarget, BuildOptions buildOptions)
     {
         string projectPath = Path.Combine(buildPath, projectName);
@@ -14,18 +13,26 @@ public static class BuildUtils
         Debug.Log("Build completed at: " + projectPath);
     }
 
-    // Method to copy folder from source to destination
     public static void CopyFolder(string sourceFolder, string destinationPath)
     {
         if (!Directory.Exists(destinationPath))
         {
             Directory.CreateDirectory(destinationPath);
+            Debug.Log($"Copied folder: {destinationPath}");
         }
 
         foreach (string file in Directory.GetFiles(sourceFolder))
         {
             string dest = Path.Combine(destinationPath, Path.GetFileName(file));
-            File.Copy(file, dest);
+            if (!File.Exists(dest))
+            {
+                File.Copy(file, dest);
+                Debug.Log($"Copied file: {file}");
+            }
+            else
+            {
+                Debug.Log($"File already exists: {file}");
+            }
         }
 
         foreach (string folder in Directory.GetDirectories(sourceFolder))
@@ -33,8 +40,6 @@ public static class BuildUtils
             string dest = Path.Combine(destinationPath, Path.GetFileName(folder));
             CopyFolder(folder, dest);
         }
-
-        Debug.Log("Folder copied to: " + destinationPath);
     }
 }
